requirements.txt
google-generativeai
python-dotenv
duckduckgo-search
streamlit
pillow
colorlog